+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [0.53.0]
 
-* Add quick fixes for some linter warnings
+* Add quick fixes for some linter errors and warnings
 * Add code navigation support for `(@)/1` goals
 * Improve readme section on code navigation